+++ baksmali/src/test/java/org/jf/baksmali/AnalysisTest.java
@@ -42,7 +42,6 @@ import org.jf.dexlib2.analysis.ClassProvider;
 import org.jf.dexlib2.iface.ClassDef;
 import org.jf.dexlib2.iface.DexFile;
 import org.jf.util.IndentingWriter;
-import org.jf.util.TextUtils;
 import org.junit.Test;
 
 import javax.annotation.Nonnull;
@@ -122,8 +121,8 @@ public class AnalysisTest {
                     className.substring(1, className.length() - 1));
             String smaliContents = readResource(smaliPath);
 
-            Assert.assertEquals(TextUtils.normalizeWhitespace(smaliContents),
-                    TextUtils.normalizeWhitespace((stringWriter.toString())));
+            Assert.assertEquals(BaksmaliTestUtils.normalizeWhitespace(smaliContents),
+                    BaksmaliTestUtils.normalizeWhitespace((stringWriter.toString())));
         }
     }
