+++ baksmali/src/test/java/org/jf/baksmali/BaksmaliTestUtils.java
@@ -33,20 +33,24 @@ package org.jf.baksmali;
 
 import com.google.common.io.ByteStreams;
 import junit.framework.Assert;
-
 import org.antlr.runtime.RecognitionException;
 import org.jf.baksmali.Adaptors.ClassDefinition;
 import org.jf.dexlib2.iface.ClassDef;
 import org.jf.smali.SmaliTestUtils;
 import org.jf.util.IndentingWriter;
-import org.jf.util.TextUtils;
+import org.junit.Test;
 
 import javax.annotation.Nonnull;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.StringWriter;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
 
 public class BaksmaliTestUtils {
+
+    private static String newline = System.getProperty("line.separator");
+
     public static void assertSmaliCompiledEquals(String source, String expected,
                                                  BaksmaliOptions options, boolean stripComments) throws IOException,
             RecognitionException {
@@ -74,9 +78,9 @@ public class BaksmaliTestUtils {
     @Nonnull
     public static String normalizeSmali(@Nonnull String smaliText, boolean stripComments) {
         if (stripComments) {
-            smaliText = TextUtils.stripComments(smaliText);
+            smaliText = stripComments(smaliText);
         }
-        return TextUtils.normalizeWhitespace(smaliText);
+        return normalizeWhitespace(smaliText);
     }
 
     @Nonnull
@@ -113,6 +117,61 @@ public class BaksmaliTestUtils {
         return new String(readResourceBytesFully(fileName), encoding);
     }
 
-    // Static helpers class; do not instantiate.
-    private BaksmaliTestUtils() { throw new AssertionError(); }
+    @Nonnull
+    public static String normalizeNewlines(@Nonnull String source) {
+        return normalizeNewlines(source, newline);
+    }
+
+    @Nonnull
+    public static String normalizeNewlines(@Nonnull String source, String newlineValue) {
+        return source.replace("\r", "").replace("\n", newlineValue);
+    }
+
+    @Nonnull
+    public static String normalizeWhitespace(@Nonnull String source) {
+        // Go to native system new lines so that ^/$ work correctly
+        source = normalizeNewlines(source);
+
+        // Remove all suffix/prefix whitespace
+        Pattern pattern = Pattern.compile("((^[ \t]+)|([ \t]+$))", Pattern.MULTILINE);
+        Matcher matcher = pattern.matcher(source);
+        source = matcher.replaceAll("");
+
+        // Remove all empty lines
+        Pattern pattern2 = Pattern.compile("^\r?\n?", Pattern.MULTILINE);
+        Matcher matcher2 = pattern2.matcher(source);
+        source = matcher2.replaceAll("");
+
+        // Remove a trailing new line, if present
+        Pattern pattern3 = Pattern.compile("\r?\n?$");
+        Matcher matcher3 = pattern3.matcher(source);
+        source = matcher3.replaceAll("");
+
+        // Go back to unix-style \n newlines
+        source = normalizeNewlines(source, "\n");
+        return source;
+    }
+
+    @Nonnull
+    public static String stripComments(@Nonnull String source) {
+        Pattern pattern = Pattern.compile("#(.*)");
+        Matcher matcher = pattern.matcher(source);
+        return matcher.replaceAll("");
+    }
+
+    @Test
+    public void testStripComments() {
+        Assert.assertEquals("", stripComments("#world"));
+        Assert.assertEquals("hello", stripComments("hello#world"));
+        Assert.assertEquals("multi\nline", stripComments("multi#hello world\nline#world"));
+    }
+
+    @Test
+    public void testNormalizeWhitespace() {
+        Assert.assertEquals("", normalizeWhitespace(" "));
+        Assert.assertEquals("hello", normalizeWhitespace("hello "));
+        Assert.assertEquals("hello", normalizeWhitespace(" hello"));
+        Assert.assertEquals("hello", normalizeWhitespace(" hello "));
+        Assert.assertEquals("hello\nworld", normalizeWhitespace("hello \n \n world"));
+    }
 }
