+++ dexlib2/build.gradle
@@ -46,7 +46,6 @@ configurations {
 }
 
 dependencies {
-    compile project(':util')
     compile depends.findbugs
     compile depends.guava
