+++ dexlib2/src/main/java/org/jf/dexlib2/dexbacked/DexBackedMethodImplementation.java
@@ -41,7 +41,7 @@ import org.jf.dexlib2.dexbacked.util.VariableSizeLookaheadIterator;
 import org.jf.dexlib2.iface.MethodImplementation;
 import org.jf.dexlib2.iface.debug.DebugItem;
 import org.jf.dexlib2.iface.instruction.Instruction;
-import org.jf.util.AlignmentUtils;
+import org.jf.dexlib2.util.AlignmentUtils;
 import org.jf.util.ExceptionWithContext;
 
 import javax.annotation.Nonnull;
