+++ dexlib2/src/main/java/org/jf/util/CharSequenceUtils.java
