+++ dexlib2/src/main/java/org/jf/util/CollectionUtils.java
