+++ dexlib2/src/main/java/org/jf/util/NumberUtils.java
