+++ dexlib2/src/main/java/org/jf/util/StringUtils.java
