+++ dexlib2/src/main/java/org/jf/util/StringWrapper.java
@@ -1,18 +1,18 @@
 /*
- * Copyright 2013, Google Inc.
+ * Copyright 2018, Google Inc.
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
  * modification, are permitted provided that the following conditions are
  * met:
  *
- *     * Redistributions of source code must retain the above copyright
+ * Redistributions of source code must retain the above copyright
  * notice, this list of conditions and the following disclaimer.
- *     * Redistributions in binary form must reproduce the above
+ * Redistributions in binary form must reproduce the above
  * copyright notice, this list of conditions and the following disclaimer
  * in the documentation and/or other materials provided with the
  * distribution.
- *     * Neither the name of Google Inc. nor the names of its
+ * Neither the name of Google Inc. nor the names of its
  * contributors may be used to endorse or promote products derived from
  * this software without specific prior written permission.
  *
@@ -185,10 +185,6 @@ public class StringWrapper {
         String[] newArr = new String[newLength];
         System.arraycopy(arr, 0, newArr, 0, arr.length);
         return newArr;
-    }
-
-    public static void printWrappedString(@Nonnull PrintStream stream, @Nonnull String string) {
-        printWrappedString(stream, string, ConsoleUtil.getConsoleWidth());
     }
 
     public static void printWrappedString(@Nonnull PrintStream stream, @Nonnull String string, int maxWidth) {
