+++ dexlib2/src/test/java/org/jf/dexlib2/writer/DexWriterSleb128Test.java
@@ -32,7 +32,6 @@
 package org.jf.dexlib2.writer;
 
 import junit.framework.Assert;
-import org.jf.util.NakedByteArrayOutputStream;
 import org.junit.Test;
 
 import java.io.IOException;
