+++ dexlib2/src/test/java/org/jf/util/NumberUtilsTest.java
