+++ smali/src/main/jflex/smaliLexer.jflex
@@ -2,6 +2,7 @@ package org.jf.smali;
 
 import java.io.*;
 import org.antlr.runtime.*;
+import org.jf.smali.util.*;
 import org.jf.util.*;
 import static org.jf.smali.smaliParser.*;
